requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr!=2.1.0,>=1.10.0 # Apache-2.0
openstacksdk==0.18.0
requests>=2.20.0
retry>=0.9.2
six>=1.11.0
systemd-python>=234
